+++ scripts/deployThalesRoyale/deprecated/deployThalesRoyale.js
@@ -4,9 +4,9 @@ const { ethers } = require('hardhat');
 
 const w3utils = require('web3-utils');
 
-const { toBytes32 } = require('../../index');
+const { toBytes32 } = require('../../../index');
 
-const { getTargetAddress, setTargetAddress, encodeCall } = require('../helpers');
+const { getTargetAddress, setTargetAddress, encodeCall } = require('../../helpers');
 
 async function main() {
